+++ content-resources/src/main/resources/flyway/oskari/V1_38_0__reset_statsgrid_default_config.sql
@@ -1,0 +1,2 @@
+-- default values have changed and are now in code, no need to use db for defaults;
+UPDATE portti_bundle set config = '{}' WHERE name = 'statsgrid';