+++ control-base/src/main/java/fi/nls/oskari/control/view/modifier/bundle/Publisher2Handler.java
@@ -1,6 +1,8 @@
 package fi.nls.oskari.control.view.modifier.bundle;
 
 import fi.nls.oskari.annotation.OskariViewModifier;
+import fi.nls.oskari.view.modifier.ModifierException;
+import fi.nls.oskari.view.modifier.ModifierParams;
 
 @OskariViewModifier("publisher2")
 public class Publisher2Handler extends PublisherHandler {