+++ control-base/src/main/java/fi/nls/oskari/control/view/modifier/bundle/PublisherHandler.java
@@ -1,7 +1,7 @@
 package fi.nls.oskari.control.view.modifier.bundle;
 
 import fi.nls.oskari.annotation.OskariViewModifier;
-import fi.nls.oskari.control.view.PublishHandler;
+import fi.nls.oskari.control.view.AppSetupHandler;
 import fi.nls.oskari.domain.Role;
 import fi.nls.oskari.log.LogFactory;
 import fi.nls.oskari.log.Logger;
@@ -14,6 +14,8 @@ import fi.nls.oskari.view.modifier.ModifierParams;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
+import java.util.Map;
+
 @OskariViewModifier("publisher")
 public class PublisherHandler extends BundleHandler {
 
@@ -21,9 +23,11 @@ public class PublisherHandler extends BundleHandler {
 
     private static JSONArray DRAW_ENABLED_ROLES = new JSONArray();
     private static final String KEY_DRAW_ROLE_IDS = "drawRoleIds";
+    private static final String KEY_TERMS_OF_USE_URL = "termsOfUseUrl";
+    public static final String PROPERTY_TERMS_OF_USE_URL = "oskari.map.terms.url";
 
     public void init() {
-        String[] roleNames = PropertyUtil.getCommaSeparatedList(PublishHandler.PROPERTY_DRAW_TOOLS_ENABLED);
+        String[] roleNames = PropertyUtil.getCommaSeparatedList(AppSetupHandler.PROPERTY_DRAW_TOOLS_ENABLED);
         try {
             final Role[] roles = UserService.getInstance().getRoles();
             for(Role role : roles) {
@@ -46,7 +50,20 @@ public class PublisherHandler extends BundleHandler {
             return false;
         }
         JSONHelper.putValue(config, KEY_DRAW_ROLE_IDS, DRAW_ENABLED_ROLES);
+        setupTerms(config);
         return false;
+    }
+    private void setupTerms(final JSONObject config) {
+        if(config.has(KEY_TERMS_OF_USE_URL)) {
+            return;
+        }
+        final Object termsObj = PropertyUtil.getLocalizableProperty(PROPERTY_TERMS_OF_USE_URL);
+        if(termsObj instanceof String) {
+            JSONHelper.putValue(config, KEY_TERMS_OF_USE_URL, termsObj);
+        } else if(termsObj instanceof Map) {
+            Map<String, String> values = (Map<String, String>) termsObj;
+            JSONHelper.putValue(config, KEY_TERMS_OF_USE_URL, new JSONObject(values));
+        }
     }
 
 }