+++ service-map/src/main/java/fi/nls/oskari/map/layer/formatters/LayerJSONFormatterWFS.java
@@ -42,6 +42,9 @@ public class LayerJSONFormatterWFS extends LayerJSONFormatter {
         }
         JSONHelper.putValue(layerJson, "isQueryable", true);
         JSONHelper.putValue(layerJson, "wps_params", getWpsParams(wfsConf) );
+        if(wfsConf != null){
+            JSONHelper.putValue(layerJson, "WMSLayerId", wfsConf.getWMSLayerId() );
+        }
 
         return layerJson;
     }