+++ service-spatineo-monitor/src/main/java/fi/nls/oskari/spatineo/dao/SpatineoMonitoringDao.java
@@ -3,79 +3,70 @@ package fi.nls.oskari.spatineo.dao;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import fi.nls.oskari.log.LogFactory;
 import fi.nls.oskari.log.Logger;
-import fi.nls.oskari.spatineo.dto.OskariMapLayerDto;
 import fi.nls.oskari.spatineo.dto.SpatineoMonitoringResponseDto;
 import org.apache.commons.httpclient.*;
-import org.apache.commons.httpclient.methods.PostMethod;
-import org.apache.commons.httpclient.params.HttpMethodParams;
 
 import java.io.IOException;
-import java.util.LinkedList;
-import java.util.List;
 
 /**
  * Data access to the Spatineo new geographical data monitoring service
  */
 public class SpatineoMonitoringDao {
 
-    private static final Logger log = LogFactory.getLogger(SpatineoServalDao.class);
+    private static final Logger log = LogFactory.getLogger(SpatineoMonitoringDao.class);
 
-    private final String serviceUrl;
+    private String serviceUrl;
     private final HttpClient httpClient;
+    private final HttpRequester httpRequester = new HttpRequester();
+    private final boolean lean = false;     // set this to true in production
 
-    private static final HttpMethodRetryHandler NO_RETRY =  new HttpMethodRetryHandler() {
+    private static final HttpMethodRetryHandler NO_RETRY = new HttpMethodRetryHandler() {
         @Override
         public boolean retryMethod(HttpMethod method, IOException exception, int executionCount) {
             return false;
         }
     };
 
-    public SpatineoMonitoringDao(final String serviceUrl, HttpClient client) {
+    /**
+     * Constructor
+     *
+     * @param client The Apache HttpClient instance to be used - needed for
+     * testing
+     * @param allMeters Include all meters regardless whether they have
+     * Indicators or not?
+     */
+    public SpatineoMonitoringDao(String serviceUrl, HttpClient client, boolean allMeters) {
         this.serviceUrl = serviceUrl;
         this.httpClient = client;
+        if (allMeters) {
+            this.serviceUrl += "&includeAllMeters=true";
+        }
+        if (lean) {
+            this.serviceUrl += "&lean=true";
+        }
     }
 
-    public SpatineoMonitoringResponseDto checkServiceStatus(final List<OskariMapLayerDto> layers) {
-        httpClient.getParams().setParameter(HttpMethodParams.RETRY_HANDLER, NO_RETRY);
-
-        if (null != System.getProperty("http.proxyHost") && null != System.getProperty("http.proxyPort")) {
-            httpClient.getHostConfiguration().setProxy(System.getProperty("http.proxyHost"),
-                    Integer.parseInt(System.getProperty("http.proxyPort")));
-        }
-
-        final PostMethod post = new PostMethod(serviceUrl);
-        
-        
-        
-        
-//        final List<NameValuePair> pairs = new LinkedList<NameValuePair>();
-//        int i = 0;
-//        for (final OskariMapLayerDto layer : layers) {
-//            pairs.add(new NameValuePair(String.format("service[%s][type]", i), "WMS"));
-//            pairs.add(new NameValuePair(String.format("service[%s][url]", i), layer.url));
-//            pairs.add(new NameValuePair(String.format("service[%s][offering]", i), layer.name));
-//            i++;
-//        }
-//        post.setRequestBody(pairs.toArray(new NameValuePair[pairs.size()]));
-        
-        
-        
-        
-        post.addRequestHeader("Origin", "http://www.paikkatietoikkuna.fi");
-        post.addRequestHeader("X-Requested-With", getClass().getName());
-        try {
-            final int status = httpClient.executeMethod(post);
-            if (200 == status) {
-                return new ObjectMapper().readValue(post.getResponseBodyAsString(), SpatineoMonitoringResponseDto.class);
-            } else {
-                return null;
+    /**
+     * This is called by SpatineoServalUpdateService class
+     * (Scheduled in portal-ext.properties, every 5min)
+     * 
+     * @return Deserialized server response 
+     */
+    public SpatineoMonitoringResponseDto checkServiceStatus() {
+        log.debug("proxyHost=" + System.getProperty("http.proxyHost") + ", proxyPort=" + System.getProperty("http.proxyPort"));
+        String response = httpRequester.GET(serviceUrl);
+        if (response != null) {
+            System.err.println("response = " + response);
+            try {
+                return new ObjectMapper().readValue(response, SpatineoMonitoringResponseDto.class);
+            } catch (IOException ex) {
+                log.debug(ex.getMessage());
+                log.debug(ex.getStackTrace());
+                for (Throwable t : ex.getSuppressed()) {
+                    log.debug(t.getMessage());
+                }
+                log.debug("Mapping Spatineo Monitoring API response to a DTO failed!");
             }
-        } catch (final HttpException e) {
-            // ignore
-        } catch (final IOException e) {
-            // ignore
-        } finally {
-            post.releaseConnection();
         }
         return null;
     }